Organize trial code
