+++ src/Api/Billing/Controllers/AccountsBillingController.cs
@@ -1,6 +1,5 @@
 ﻿#nullable enable
 using Bit.Api.Billing.Models.Responses;
-using Bit.Core.Billing.Models.Api.Requests.Accounts;
 using Bit.Core.Billing.Payment.Services;
 using Bit.Core.Billing.Premium.Requests;
 using Bit.Core.Billing.Services;
