+++ src/SharedWeb/Utilities/ServiceCollectionExtensions.cs
@@ -21,9 +21,7 @@ using Bit.Core.Auth.Services;
 using Bit.Core.Auth.Services.Implementations;
 using Bit.Core.Auth.UserFeatures;
 using Bit.Core.Billing.Payment.Services;
-using Bit.Core.Billing.Services;
-using Bit.Core.Billing.Services.Implementations;
-using Bit.Core.Billing.TrialInitiation;
+using Bit.Core.Billing.Trials;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.HostedServices;
